+++ internal/align/output_test.go
@@ -93,4 +93,3 @@ func TestOutputBlockOrder(t *testing.T) {
 }`
 	require.Equal(t, exp, got)
 }
-