+++ internal/align/strategy.go
@@ -4,7 +4,7 @@ package align
 import "github.com/hashicorp/hcl/v2/hclwrite"
 
 type Options struct {
-	Order       []string
+	Order []string
 	PrefixOrder bool
 
 	Schemas map[string]*Schema